argon2/pwhash: Begin updating high-level pwhash
