+++ tests/kdf/custom_argon2.rs
@@ -83,6 +83,7 @@ fn run_tests_from_json(path_to_vectors: &str) {
                 )
                 .unwrap();
                 assert_eq!(rt.unprotected_as_ref(), test.phc.as_bytes());
+                assert_eq!(rt.unprotected_as_str(), test.phc.as_str());
                 assert!(
                     Argon2::<I, Sequential>::verify_encoded(
                         &phc,
@@ -121,6 +122,7 @@ fn run_tests_from_json(path_to_vectors: &str) {
                 )
                 .unwrap();
                 assert_eq!(rt.unprotected_as_ref(), test.phc.as_bytes());
+                assert_eq!(rt.unprotected_as_str(), test.phc.as_str());
                 assert!(
                     Argon2::<ID, Sequential>::verify_encoded(
                         &phc,
